Change unsigned integers to signed double for data transmission

+++ Arduino/IIT_HASP_Payload/Comm.h
@@ -26,7 +26,7 @@
 /*******************************/
 
 void COMM_setup();
-void COMM_sendSensors(unsigned long* sensorArray, unsigned long time);
+void COMM_sendSensors(double *sensorArray, unsigned long time);
 void COMM_readSerial();
 
 #endif
